Create Note.applescript
Use this AppleScript code in the Automator app.
Opens the Stickies app if it's closed. Creates a new Sticky Note if it's already open.
Opens purple notes. 
In the Automator app, add this Applescript as a part of the workflow. Assign it to the function key.
If you have a touch bar, style the button however you want.