Create Time.applescript
Use this AppleScript code in the Automator app.
It enters the current time in "hh:mm <AM/PM>" format, wherever the cursor is. 
In the Automator app, add this Applescript as a part of the workflow. Assign it to the function key.
If you have a touch bar, style the button however you want.